requirements.txt
ipykernel
pandas
numpy
matplotlib
plotly
seaborn
scikit-learn
pymongo
from_root
dill
certifi
PyYAML
mypy-boto3-s3
botocore
fastapi
python-multipart
uvicorn
jinja2
imblearn
azure.storage.blob
-e.